Reorg to be able to use structured logging in shared files.
Using arguments.cc as a test-case.
