+++ pdns/rec_control.cc
@@ -85,8 +85,10 @@ static void initArguments(int argc, char** argv)
   }
 }
 
+
 int main(int argc, char** argv)
 {
+  g_slogStructured = false;
   const set<string> fileCommands = {
     "dump-cache",
     "dump-edns",
